removed old api gen files
