--- translator/Example.cpp
@@ -1,13 +1,0 @@
-#include "TestOut.hpp"
-
-CANHelper::CanMsgHandler CAN;
-
-void setup()
-{
-
-}
-
-void loop()
-{
-
-}
